Add solution to first part
The solution to first part is a really simple solution where you input
the "calories" in a text file and iterates over the input file. We then
check each line and validate that it is an integer. If it is an integer
and we add it to the "elfs total calories" if it is not an integer we
compare it to the "most calories" value and update it if the elfs
calories are higher.